// src/schemas/track.ts
import { z } from 'zod';

const TrackType = z.object({
  track_type: z.string()
});

export const Track = z.object({
  ai_enabled: z.boolean(),
  allow_pitlane_collisions: z.boolean(),
  allow_rolling_start: z.boolean(),
  allow_standing_start: z.boolean(),
  award_exempt: z.boolean(),
  category: z.string(),
  category_id: z.number(),
  closes: z.string(),
  config_name: z.string(),
  corners_per_lap: z.number(),
  created: z.string(),
  first_sale: z.string(),
  free_with_subscription: z.boolean(),
  fully_lit: z.boolean(),
  grid_stalls: z.number(),
  has_opt_path: z.boolean(),
  has_short_parade_lap: z.boolean(),
  has_start_zone: z.boolean(),
  has_svg_map: z.boolean(),
  is_dirt: z.boolean(),
  is_oval: z.boolean(),
  is_ps_purchasable: z.boolean(),
  lap_scoring: z.number(),
  latitude: z.number(),
  location: z.string(),
  longitude: z.number(),
  max_cars: z.number(),
  night_lighting: z.boolean(),
  nominal_lap_time: z.number(),
  number_pitstalls: z.number(),
  opens: z.string(),
  package_id: z.number(),
  pit_road_speed_limit: z.number(),
  price: z.number(),
  price_display: z.string(),
  priority: z.number(),
  purchasable: z.boolean(),
  qualify_laps: z.number(),
  rain_enabled: z.boolean(),
  restart_on_left: z.boolean(),
  retired: z.boolean(),
  search_filters: z.string(),
  site_url: z.string(),
  sku: z.number(),
  solo_laps: z.number(),
  start_on_left: z.boolean(),
  supports_grip_compound: z.boolean(),
  tech_track: z.boolean(),
  time_zone: z.string(),
  track_config_length: z.number(),
  track_dirpath: z.string(),
  track_id: z.number(),
  track_name: z.string(),
  track_types: z.array(TrackType)
});

export const Tracks = z.array(Track);